Update Dockerfile to fix issues with debian-slim
An update to debian:latest-slim seems to have removed some
packages that we were depending on (JDK 11 and some fonts).

This new approach targets a fixed Dockerfile target (instead of
"latest"), so it should be more future-proof. Also, it lets us
delete some code, which is always a win when you can get away
with it!